collector/requirements.txt
feedparser==6.0.11
requests==2.31.0
google-api-python-client==2.108.0
supabase==2.3.0
python-dotenv==1.0.0
